requirements.txt
psycopg2
langchain
openai
tiktoken
python-dotenv
